workshop2/streamlit/requirements.txt
streamlit
mysql-connector-python
requests
langchain
langchain-community
pyautogen
python-dotenv
ag2[openai]
google-generativeai
vertexai
ag2[gemini]
